Setting turn_carousel_state based on 'Share'
Raw actions are split by messages, not necissarily by when sharing
occurs. This is especially noticible when, say, a SpecifyInfo action is
inserted which doesn't often align with a share actions. Thus need to
take into acount when 'Share' occurs to determine the final scene as seen
by the user as the result of an one round.